Fixed warning messages from VC++ Added missing diagram for slider documentation (hopefully I figured out how to tell CVS that it is binary, I used "-kb")
git-svn-id: file:///fltk/svn/trunk@2203 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fltk/Fl_Valuator.h
@@ -1,5 +1,5 @@
 //
-// "$Id: Fl_Valuator.h,v 1.5 2002/05/06 06:31:26 spitzak Exp $"
+// "$Id: Fl_Valuator.h,v 1.6 2002/05/07 04:58:19 spitzak Exp $"
 //
 // Valuator header file for the Fast Light Tool Kit (FLTK).
 //
@@ -38,7 +38,7 @@ public:
   int value(double);
 
   float step() const {return step_;}
-  void step(float a) {step_ = a;}
+  void step(double a) {step_ = float(a);}
 
   double minimum() const {return minimum_;}
   void minimum(double a) {minimum_ = a;}
@@ -95,5 +95,5 @@ private:
 #endif
 
 //
-// End of "$Id: Fl_Valuator.h,v 1.5 2002/05/06 06:31:26 spitzak Exp $".
+// End of "$Id: Fl_Valuator.h,v 1.6 2002/05/07 04:58:19 spitzak Exp $".
 //
